add stats

+++ Assets/Scripts/Gameplay/Combat/Damage/DamageInfo.cs
@@ -4,12 +4,12 @@ namespace LayerZero.Gameplay.Combat.Damage
 {
     internal readonly struct DamageInfo
     {
-        public readonly int Amount;
+        public readonly float Amount;
         public readonly DamageSource Source;
         public readonly Transform AttackerTransform;
         public readonly DamageImpactInfo Impact;
 
-        public DamageInfo(int amount, DamageSource source, Transform attackerTransform, DamageImpactInfo impact = default)
+        public DamageInfo(float amount, DamageSource source, Transform attackerTransform, DamageImpactInfo impact = default)
         {
             Amount = amount;
             Source = source;
